requirements.txt
alembic==1.9.2
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
fastapi-cache2==0.2.0
fastapi-users==10.3.0
fastapi-users-db-sqlalchemy==4.0.4
flower==1.2.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanize==4.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
makefun==1.15.0
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.5
passlib==1.7.4
pendulum==2.1.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0
redis==4.4.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tornado==6.2
typing_extensions==4.11.0
ujson==5.7.0
uvicorn==0.20.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
ultralytics~=8.2.0
firebase-admin==6.5.0
pillow~=10.3.0
torch~=2.2.2
numpy~=1.26.4
flower
gunicorn
opencv-python
